requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.28.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Social sentiment - ADVANCED
beautifulsoup4>=4.11.0
feedparser>=6.0.0
vaderSentiment>=3.3.2
textblob>=0.17.1
tweepy>=4.14.0
snscrape>=0.4.3.20220106
praw>=7.6.0

# Financial data APIs
alpha_vantage>=2.3.1
finnhub-python>=2.4.19
quandl>=3.7.0
polygon-api-client>=1.12.0

# AI Integration
openai>=1.3.0
anthropic>=0.8.0
transformers>=4.35.0
torch>=2.0.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.9.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Real-time processing
kafka-python>=2.0.2
redis>=4.5.0
websocket-client>=1.6.0

# Visualization & Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.14.0

# Utilities
python-dotenv>=0.19.0
schedule>=1.2.0
pytz>=2023.3
dateutil>=2.8.2